Add verifier script for deploying vested wallets
