+++ package.json
@@ -15,6 +15,7 @@
     "deploy:governance": "ts-node ./scripts/governanceDeployer.ts",
     "deploy:note-store": "ts-node ./scripts/deployNoteStore.ts",
     "deploy:vested-wallets": "ts-node ./scripts/vestedWalletsDeployer.ts",
+    "verify:vested-wallets": "ts-node ./scripts/vestedWalletsDeploymentVerifier.ts",
     "verify:governance": "ts-node ./scripts/governanceDeploymentVerifier.ts",
     "allocate:tokens": "ts-node ./scripts/tokenAllocator.ts",
     "allocate:dao:tokens": "ts-node ./scripts/tokenDaoAllocator.ts",
